worker/requirements.txt
hatchet-sdk
supabase
python-dotenv
pydantic
